* Remove DPLA id generation from OAI, API and RS harvesters. This ID generation will happen during mapping so the harvesters can be utilized outside DPLA. * Moved #getOaiIdentifier to OaiResponseProcessor * Deleted Harvester and HarvesterTest * Moved and renamed HarvesterException case class to OaiResponseProcessor as OaiHarvesterException
